1. Finished the method for foliations that calculates the transition data for restricting to an interval.
2. The lengths of the Arnoux-Yoccoz foliation were in QQbar, because the
stretch factor is calculated of an eigenvalue of a matrix over QQ. This
slowed down the calculation by at least an order of magnitude. This is
now fixed and the calculation is done by floating point numbers.

3. Added a separate _fmod method on top of _mod, because there was an
issue with floating point division. It would be nice later to keep only
one of these.