There has been a lot of not committed changes recently.
The older, but still working versions of the algorithm and the
involution class were moved into old_algorithm.sage and
old_involution.sage. They are not intended to be changed in the future
and kept only for safety and reconstruction purposes.

polynomial_filtering.sage is a version of Jean-Luc's and Erwan's
algorithm, implemented to work for non-orientable surfaces, too, though
it is somewhat less sophisticated then their Mathematica implementation.

involution.sage is not nicely doctested, simplified and made more
elegant at certain places.

foliation.sage uses the Involution class in involution.sage, and since
doctesting has problems with importing from another file, I had to copy
the contents of involution.sage into foliation.sage in order to doctest
foliation.sage. Doctesting has begun in foliation.sage as well, and now
I am planning to unify the orientable and non-orientable cases more to
simplify the code. The PointWithCoefficients class will become much
simpler along with other parts of the code, including the Foliation
class, but other parts might become a little more complicated, but
hopefully it will work out. I am doing this commit to be able to revert
in case it becomes totally messed up.